+++ library/src/com/actionbarsherlock/internal/widget/ActionBarView.java
@@ -91,7 +91,7 @@ public final class ActionBarView extends RelativeLayout {
 
     public ActionBarView(Context context, AttributeSet attrs, int defStyle) {
         super(context, attrs, defStyle);
-        LayoutInflater.from(context).inflate(R.layout.action_bar, this, true);
+        LayoutInflater.from(context).inflate(R.layout.abs__action_bar, this, true);
 
         final TypedArray attrsActionBar = context.obtainStyledAttributes(attrs, R.styleable.SherlockTheme, defStyle, 0);
         mIsConstructing = true;
@@ -532,7 +532,7 @@ public final class ActionBarView extends RelativeLayout {
     }
 
     public ActionBarView.Item newItem() {
-        ActionItem item = (ActionItem)LayoutInflater.from(getContext()).inflate(R.layout.action_bar_item_layout, mActionsView, false);
+        ActionItem item = (ActionItem)LayoutInflater.from(getContext()).inflate(R.layout.abs__action_bar_item_layout, mActionsView, false);
         item.setActionBar(this);
         return item;
     }
@@ -762,7 +762,7 @@ public final class ActionBarView extends RelativeLayout {
 
         public HomeItem(Context context, AttributeSet attrs, int defStyle) {
             super(context, attrs, defStyle);
-            LayoutInflater.from(context).inflate(R.layout.action_bar_home, this, true);
+            LayoutInflater.from(context).inflate(R.layout.abs__action_bar_home, this, true);
 
             mLogo = (ImageView)findViewById(R.id.actionbarwatson_home_logo);
             mIcon = (ImageView)findViewById(R.id.actionbarwatson_home_icon);
@@ -877,7 +877,7 @@ public final class ActionBarView extends RelativeLayout {
 
         TabImpl(ActionBarView actionBar) {
             mActionBar = actionBar;
-            mView = LayoutInflater.from(mActionBar.getContext()).inflate(R.layout.action_bar_tab_layout, actionBar.mTabsView, false);
+            mView = LayoutInflater.from(mActionBar.getContext()).inflate(R.layout.abs__action_bar_tab_layout, actionBar.mTabsView, false);
             mView.setTag(this);
             mView.setOnClickListener(clickListener);