+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Version 3.1.3 *(In Development)*
 
 * Fix: Action bar background being set on two views causing artifacts to remain
   on screen when the action bar was hidden.
+* Fix: Incorrect sub-menu item being selected by default when the sub-menu was
+  triggered from the native options menu on pre-3.0.
 
 
 Version 3.1.2 *(2011-08-07)*